# apps/desktop-companion/openapi.yaml
openapi: 3.1.0
info:
  title: FluxBoard Livestream Copilot API
  description: |
    REST API for the FluxBoard Livestream Copilot desktop companion service.

    FluxBoard provides real-time livestream capture, AI-powered content generation,
    OBS Studio integration, clip management, social media export, and video streaming
    for content creators.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Bearer JWT** - Pass an access token in the `Authorization` header.
    2. **API Key** - Pass an API key in the `x-api-key` header.

    Public endpoints (health checks, OBS status, event reads) do not require authentication
    during development.

    ## WebSocket Events

    The server exposes a WebSocket on a configurable port (default 9001). Clients receive
    real-time `EventEnvelope` messages for session lifecycle, transcripts, clips, outputs,
    and trigger detections. Connect and listen for JSON messages.

    Event types: `SESSION_START`, `SESSION_END`, `TRANSCRIPT_SEGMENT`, `MOMENT_MARKER`,
    `CLIP_INTENT_START`, `CLIP_INTENT_END`, `ARTIFACT_CLIP_CREATED`, `ARTIFACT_FRAME_CREATED`,
    `OUTPUT_CREATED`, `OUTPUT_VALIDATED`, `AUTO_TRIGGER_DETECTED`, `CLIP_QUEUE_UPDATED`.

    ## Rate Limiting

    Rate-limited endpoints return standard `RateLimit-*` headers. When exceeded, the server
    responds with `429 Too Many Requests`.

    | Endpoint Group        | Limit              |
...

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login or /api/v1/auth/refresh
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Static API key for service-to-service communication
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: integer
                      tokenType:
                        type: string
                        example: Bearer
        "401":
          description: Invalid or expired refresh token
...
  /api/v1/auth/oauth/{provider}:
    get:
      tags: [OAuth]
      summary: Initiate OAuth flow (redirects to provider)
      operationId: initiateOAuth
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, twitch]
      responses:
        "302":
          description: Redirect to OAuth provider
        "400":
          description: Invalid or unconfigured provider
    delete:
      tags: [OAuth]
      summary: Unlink OAuth account
      operationId: unlinkOAuth
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, twitch]
      responses:
        "200":
          description: Account unlinked
        "400":
          description: Cannot unlink last auth method
  /api/v1/auth/oauth/{provider}/callback:
    get:
      tags: [OAuth]
      summary: OAuth callback handler (redirects to frontend)
      operationId: oauthCallback
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to frontend with tokens
  /api/v1/auth/oauth/{provider}/link:
    post:
      tags: [OAuth]
      summary: Link OAuth account to current user
      operationId: linkOAuth
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, twitch]
      responses:
        "200":
          description: Authorization URL for linking
        "400":
          description: Already linked
...